backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# AI/ML Libraries
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1

# PDF Processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional: For better performance
accelerate==0.25.0
